Revert "rename `getAvatar.ts` -> `getEnsAvatar.ts`"
This reverts commit 098d36803c48b5d7b5603d0d3fa5af0607126eef.
