+++ src/identity/core/getAvatar.ts
