+++ src/identity/index.ts
@@ -2,7 +2,7 @@
 export { Avatar } from './components/Avatar';
 export { Badge } from './components/Badge';
 export { Name } from './components/Name';
-export { getAvatar, getEnsAvatar } from './core/getEnsAvatar';
+export { getAvatar, getEnsAvatar } from './core/getAvatar';
 export { getName } from './core/getName';
 export { useAvatar } from './hooks/useAvatar';
 export { useName } from './hooks/useName';
